Basic level
Ram and Rohan are participating in a tournament. Together they must face off against x contenders whose strengths are given in an array arr[]. The respective strength of Ram and Rohan in m and n. They take turns taking part in face-off matches with contenders and Ram goes first. They win a match only if their opponent's strength is an integral multiple of their strengths. If the opponent is defeated, he can not fight again.

Find out who among Ram and Rohan is a better player. If both of them win the same number of matches then both are winners.